fix: unnecessary type casting

+++ apps/laboratory/src/components/Solana/SolanaSignMessageTest.tsx
@@ -21,7 +21,7 @@ export function SolanaSignMessageTest() {
 
       toast({
         title: ConstantsUtil.SigningSucceededToastTitle,
-        description: signature as Uint8Array,
+        description: signature,
         type: 'success'
       })
     } catch (err) {
